% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-introspection.R
\name{print.tce_fields_tbl}
\alias{print.tce_fields_tbl}
\title{Print method for \code{tce_fields()} output}
\usage{
\method{print}{tce_fields_tbl}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=tce_fields]{tce_fields()}}.}

\item{...}{Passed to the next print method.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{tce_fields()} output
}
\keyword{internal}
