% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remessa.R
\name{tce_benefit_types}
\alias{tce_benefit_types}
\title{Benefit types (Remessa)}
\usage{
tce_benefit_types(
  ...,
  verbose = getOption("tceper.verbose", FALSE),
  cache = TRUE,
  clean_names = TRUE
)
}
\arguments{
\item{...}{Named query parameters passed directly to the API.}

\item{verbose}{If \code{TRUE}, prints the final API URL.}

\item{cache}{If \code{TRUE} (default), returns cached results when available.}

\item{clean_names}{If \code{TRUE} (default), converts column names to \code{snake_case}.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Lookup table of benefit types (RemessaTCEPE system).
}
\seealso{
Other remessa: 
\code{\link{tce_budget_statuses}()},
\code{\link{tce_legal_instrument_documents}()},
\code{\link{tce_legal_instrument_items}()},
\code{\link{tce_legal_instrument_participants}()},
\code{\link{tce_legal_instruments}()},
\code{\link{tce_procurement_process_budget}()},
\code{\link{tce_procurement_process_documents}()},
\code{\link{tce_procurement_process_participants}()},
\code{\link{tce_procurement_processes}()},
\code{\link{tce_reference_codes}()},
\code{\link{tce_reference_dates}()},
\code{\link{tce_reference_sources}()},
\code{\link{tce_remessa_works}()},
\code{\link{tce_remessa_works_execution}()},
\code{\link{tce_remessa_works_geometry}()},
\code{\link{tce_update_indices}()}
}
\concept{remessa}
