% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introspection.R
\name{tce_catalog}
\alias{tce_catalog}
\title{List available endpoints}
\usage{
tce_catalog(search = NULL)
}
\arguments{
\item{search}{Optional text to filter by endpoint name/description.}
}
\value{
A tibble with endpoint names and basic metadata.
}
\description{
List available endpoints
}
