% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{tce_inactivation_reasons}
\alias{tce_inactivation_reasons}
\title{Inactivation reason types}
\usage{
tce_inactivation_reasons(
  ...,
  verbose = getOption("tceper.verbose", FALSE),
  cache = TRUE,
  clean_names = TRUE
)
}
\arguments{
\item{...}{Named query parameters passed directly to the API.}

\item{verbose}{If \code{TRUE}, prints the final API URL.}

\item{cache}{If \code{TRUE} (default), returns cached results when available.}

\item{clean_names}{If \code{TRUE} (default), converts column names to \code{snake_case}.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Lookup table of inactivation reason types.
}
\seealso{
Other reference: 
\code{\link{tce_creditor_types}()},
\code{\link{tce_entities}()},
\code{\link{tce_funding_sources}()},
\code{\link{tce_municipal_entities}()},
\code{\link{tce_municipalities}()},
\code{\link{tce_payroll_types}()},
\code{\link{tce_state_entities}()},
\code{\link{tce_sub_units}()}
}
\concept{reference}
