% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tceper-package.R
\docType{package}
\name{tceper-package}
\alias{tceper}
\alias{tceper-package}
\title{tceper: Access the Open Data 'API' of TCE-PE (Pernambuco Court of Accounts)}
\description{
An R interface to the Open Data 'API' of the Tribunal de Contas do Estado de Pernambuco (TCE-PE), the Court of Accounts of the State of Pernambuco, Brazil. Provides tidy, ready-to-use functions to query public data on revenues, expenditures, commitments, procurement, contracts, agreements, public works, legal processes, personnel and reference tables for all state and municipal government entities in Pernambuco. All results are returned as tibbles with column names converted to 'snake_case' by default. Uses 'httr2' for 'HTTP' requests and 'cli' for user-friendly messages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StrategicProjects/tceper}
  \item \url{https://seu-usuario.github.io/tceper/}
  \item Report bugs at \url{https://github.com/StrategicProjects/tceper/issues}
}

}
\author{
\strong{Maintainer}: Andre Leite \email{leite@castlab.org}

Authors:
\itemize{
  \item Marcos Wasilew \email{marcos.wasilew@gmail.com}
  \item Hugo Vasconcelos \email{hugo.vasconcelos@ufpe.br}
  \item Carlos Amorin \email{carlos.agaf@ufpe.br}
  \item Diogo Bezerra \email{diogo.bezerra@ufpe.br}
}

}
\keyword{internal}
